+++ src/main/java/seedu/address/model/pdfmanager/pdftasksummary/PdfWrapperLayout.java
@@ -1,4 +1,4 @@
-package seedu.address.model.pdfmanager;
+package seedu.address.model.pdfmanager.pdftasksummary;
 
 import java.time.LocalDate;
 import java.util.List;
